Don't run cronjob when no DATABASE_URL is defined (#594)

+++ src/utils/backend/worker/data-worker.ts
@@ -117,6 +117,8 @@ defineCronJob('20 */2 * * *', async () => {
 });
 
 defineCronJob('*/15 * * * *', async () => {
+    if (!process.env.DATABASE_URL) return;
+
     await prisma.userToken.deleteMany({
         where: {
             refreshMaxDate: {
